Move hidden layer configuration
Move config out of the types file,
into route utils, where the last of
the network/layer/scope configuration lives.
