+++ src/app/components/LayerPicker/LayerMenu.tsx
@@ -9,9 +9,9 @@ import Tooltip from '@mui/material/Tooltip'
 import { COLORS } from '../../../styles/theme/colors'
 import { Layer } from '../../../oasis-nexus/api'
 import { getLayerLabels } from '../../utils/content'
-import { RouteUtils } from '../../utils/route-utils'
+import { isLayerHidden, RouteUtils } from '../../utils/route-utils'
 import { Network } from '../../../types/network'
-import { isLayerHidden, orderByLayer } from '../../../types/layers'
+import { orderByLayer } from '../../../types/layers'
 import { useScreenSize } from '../../hooks/useScreensize'
 import { useScopeParam } from '../../hooks/useScopeParam'
