+++ src/app/pages/ConsensusDashboardPage/ParaTimesCard.tsx
@@ -6,10 +6,9 @@ import CardHeader from '@mui/material/CardHeader'
 import CardContent from '@mui/material/CardContent'
 import { styled } from '@mui/material/styles'
 import { Layer, Runtime } from '../../../oasis-nexus/api'
-import { isNotOnHiddenLayer } from '../../../types/layers'
 import { COLORS } from '../../../styles/theme/colors'
 import { CardHeaderWithCounter } from '../../components/CardHeaderWithCounter'
-import { RouteUtils } from '../../utils/route-utils'
+import { isNotOnHiddenLayer, RouteUtils } from '../../utils/route-utils'
 import { SearchScope } from '../../../types/searchScope'
 import { EnabledRuntimePreview, InactiveRuntimePreview } from './RuntimePreview'
